+++ src/main/java/gtclassic/util/jei/category/GTJeiMultiRecipeCategory.java
@@ -33,7 +33,9 @@ public class GTJeiMultiRecipeCategory implements IRecipeCategory<GTJeiMultiRecip
 	}
 
 	private int getHeight() {
-		return GTConfig.debugMode ? 100 : 90;
+		int baseHeight = this.name == "gt.fusion" ? 100 : 90;
+		int extraHeight = GTConfig.debugMode ? 10 : 0;
+		return baseHeight + extraHeight;
 	}
 
 	@Override